authorization_service/requirements.txt
python-dateutil
pycryptodome
pyopenssl
fpdf
redis
flask
flask-jwt-extended
flask-cors
